be/requirements.txt
flask==2.3.3
pdf2image==1.17.0
pillow==11.2.1
requests==2.32.3
python-dotenv==1.0.1
openai==1.30.1
